+++ test/naiv_standard_test.go
@@ -1,13 +1,12 @@
-package algoritmos_test
+package test
 
 import (
-	"fmt"
 	"generador/algoritmos"
 	"reflect"
 	"testing"
 )
 
-func TestNaivLoopUnrollingFour(t *testing.T) {
+func TestNaivStandard(t *testing.T) {
 	a := [][]int{
 		{1, 2, 3},
 		{4, 5, 6},
@@ -24,11 +23,9 @@ func TestNaivLoopUnrollingFour(t *testing.T) {
 		{139, 154},
 	}
 
-	resultado := algoritmos.NaivLoopUnrollingFour(a, b)
-
-	fmt.Println(resultado)
+	resultado := algoritmos.NaivStandard(a, b)
 
 	if !reflect.DeepEqual(resultado, esperado) {
-		t.Error("NaivLoopUnrollingFour ha fallado")
+		t.Error("NaivStandard ha fallado")
 	}
 }
